# Use asyncio and run LM in batches