## Background

For each example we evaluate, I expect our code to be organized according to the following steps:

* Preprocessing: generate language model (LM) inputs in whatever way is appropriate for the task
* Model inference: run the LM, possibly multiple times
* Postprocessing: take the LM's output, and score it for accuracy or the like

But if we just naively do that for each example, we run into a serious problem. If we run the examples one by one, we will drastically underutilize our GPUs, or whatever other accelerators we use.

To keep our GPUs busy, we need to process LM inputs in batches. Whatever batching system we go with, we will need to carry out the following tasks:
* Gathering LM requests and queueing them up for inference
* Taking a batch of requests and actually running the LM on them
* Following up on the LM's outputs

## This PR's changes

This PR introduces `asyncio`, and uses it to centralize all LM inference in an `InferenceWorker` class. With `asyncio`, you can write code for each evaluation task without having to worry about batching. Instead, you can call `InferenceWorker` using its one-example-at-a-time interface, and `InferenceWorker` will manage the batching internally.

This PR also makes `InferenceWorker` actually do batching, and so model evaluation is now much faster.

## Performance and accuracy

Since this PR is purely a refactor and performance improvement, the StoryCloze accuracy shouldn't change. I measured the accuracy with and without this PR, and everything looked okay!

Code | Model | Batch size | GPU | Wall-clock time (after model is loaded) | Accuracy on StoryCloze
-- | -- | -- | -- | -- | --
Before | GPT-2-117M | 1 | T4 | ~65 s | 52.4%
With PR | GPT-2-117M | 1 | T4 | ~65 s | 52.4%
With PR | GPT-2-117M | 1 | V100 | ~65 s | 52.4%
With PR | GPT-2-117M | 10 | V100 | 9.2 s | 52.4%
With PR | GPT-2-117M | 20 | V100 | 7.1 s | 52.4%
With PR | GPT-2-774M | 20 | V100 | 38.2 s | 66.1%

## This PR's programming model

* To implement an evaluation task, you can write a single function that operates on a single example. The function will need to be async, and it will look something like this:

```
async def eval_for_my_task(worker, example):
    prompt1 = ...
    prompt2 = ...

    likelihood1 = await worker.run(...)
    likelihood2 = await worker.run(...)

    return evaluate(likelihood1, likehood2, example.actual_answer)
```

* To support a new model, you can extend `InferenceWorker` 

## Alternative, if we don't want to use `asyncio`

I think our second best option is to make each evaluation task implement its own batching logic. For example, for StoryCloze, the code would look something like this:

```
def eval_storycloze_batch(lm, batch_of_examples):
    storycloze_prompts_for_sentence_1 = [generate_storycloze_prompt(example) for example in batch_of_examples]
    storycloze_prompts_for_sentence_2 = [generate_storycloze_prompt(example) for example in batch_of_examples]

    likelihoods_for_sentence_1 = lm.loglikelihood_batched(storycloze_prompts_for_sentence_1)
    likelihoods_for_sentence_2 = lm.loglikelihood_batched(storycloze_prompts_for_sentence_2)

    batch_eval_results = []
    for example, sentence_1_likelihood, sentence_2_likelihood in zip(...):
        batch_eval_results.append(...)

    return batch_eval_results

def main():
    # - find a reasonable batch size (which may be model-dependent and need to be configured)
    # - load the dataset and split the examples into matches
    # - call eval_storycloze_batch on each batch
    # - aggregate the results and compute the overall accuracy
```