@uyhcire 2020-12-27
As discussed with @leogao2, I'll abandoning asyncio, and we'll make each evaluation task responsible for collecting model inputs into batches